Updated common functions to consolidate capabilities. Put Hashcode generation, GUIDs, and regular expressions generation
