--- common/generate_hashcode.py
@@ -1,13 +1,0 @@
-import random
-import hashlib
-
-def generate_hashcode( attrib1: str,attrib2: str,attrib3: str,
-                       attrib4: str,attrib5: str,attrib6: str,
-                       attrib7: str,attrib8: str,attrib9: str)-> str:
-    hash_dataset=""
-    hash_value = hashlib.sha512(hash_dataset.encode('utf-8')).hexdigest()
-    return hash_value
-
-if __name__ == '__main__':
-    generate_hashcode(attrib1="",attrib2="",attrib3="",attrib4="",attrib5="",
-                      attrib6="",attrib7="",attrib8="",attrib9="")
